feat: support permit2
Permit2 allows us to make a single approval and then use gas-free signatures to approve transactions from then on.
This is now supported for vToken trades, creating liquidity positions, and increasing a liquidity position.
For the latter 2 methods, this is an opt-in and you can choose to just use regular approvals instead.