fix: calculate pool revenue
The daily/weekly pool revenue was showing as 0 up until now.
This was first due to it being in USD terms (and lacking a stable USD pair on testnet),
but also because the revenue values were based purely on the AMM pools, whereas most of NFTX
is based around vault fees rather than AMM fees.
So for the 0.3% pool we're now going to calculate the revenue by looking at the vault fees subgraph.