fix(@nftx/subgraph): rework and/or operators to work together nicely
we recently added support for OR operators, however if you use an AND operator in the same clause, it would not work.
This fix does some additional work to massage the output into nested and/or clauses in order to create a valid query.

As part of this fix the createQuery method has been refactored to be easier to understand and maintain.